added `ItemSerializerTest.php` instead of former `ItemRepositoryTest.php`
